requirements.txt
async-timeout==4.0.2
bidict==0.22.0
click==8.1.3
dateutils==0.6.12
Deprecated==1.2.13
Flask==2.1.3
Flask-SocketIO==5.2.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.3
python-socketio==5.7.1
pytz==2022.1
redis==4.3.4
six==1.16.0
Werkzeug==2.2.0
wrapt==1.14.1
zipp==3.8.1
